Decouple using fetched url from fetch function
Fetch function only returns the image URL and it should not do any more
than that, assigning url to the image source for example.

Returning a settled promise, the fetch function gets called within the
`updateImage` function, where it's chained with a `then` method that
updates the image source with the result url.